refactor: update webapp/ui and scoops imports to kernel/messages.ts
Redirect UI, boot, and scoop modules from the cross-package
chrome-extension path to the webapp-local kernel/messages.js.

Part of #1443
