+++ packages/webapp/src/ui/runtime-mode.ts
@@ -1,5 +1,5 @@
 import { ELECTRON_OVERLAY_APP_PATH } from '@slicc/shared-ts';
-import { DETACHED_RUNTIME_QUERY_NAME } from '../../../chrome-extension/src/messages.js';
+import { DETACHED_RUNTIME_QUERY_NAME } from '../kernel/messages.js';
 import {
   type RuntimeConfigStorage,
   resolveFollowerJoinUrl,
@@ -21,7 +21,7 @@ export const HOSTED_LEADER_RUNTIME_QUERY_VALUE = 'hosted-leader';
 // Re-export shared detached-runtime constants from chrome-extension/messages.ts
 // so panel-side code (resolveUiRuntimeMode) and SW-side code share the same
 // source of truth.
-export { DETACHED_RUNTIME_QUERY_NAME } from '../../../chrome-extension/src/messages.js';
+export { DETACHED_RUNTIME_QUERY_NAME } from '../kernel/messages.js';
 
 export function resolveUiRuntimeMode(
   locationHref: string,
